refactor: nft page components
